# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "CEP API",
        "title": "Cep API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Guilherme Daniel",
            "url": "https://github.com/Guilherme415"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/cep/{cep}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Api to get address details by cep. This Api try to get the infos in one or more APIs and return the fastest\nIf the Cep is invalid, the api will replace the lasts digits to 0 until find some valid Cep\nIf does not find some valid cep, it will return an internal server error",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cep"
                ],
                "summary": "Get address details by CEP",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CEP",
                        "name": "cep",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.GetAddressDeitalsByCEPResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "response.GetAddressDeitalsByCEPResponse": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "neighborhood": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}